Update cann versions and docs

+++ examples/01_point_to_point/02_batch_send_recv_ring/main.cc
@@ -38,7 +38,7 @@
 
 struct ThreadContext {
     HcclRootInfo *rootInfo;
-    int32_t device;
+    uint32_t device;
     uint32_t devCount;
 };
 
@@ -48,12 +48,12 @@ int Sample(void *arg)
     void *hostBuf = nullptr;
     void *sendBuf = nullptr;
     void *recvBuf = nullptr;
-    int32_t device = ctx->device;
-    uint32_t count = ctx->devCount;
+    uint32_t device = ctx->device;
+    uint64_t count = ctx->devCount;
     size_t mallocSize = count * sizeof(float);
 
     // 设置当前线程操作的设备
-    ACLCHECK(aclrtSetDevice(device));
+    ACLCHECK(aclrtSetDevice(static_cast<int32_t>(device)));
 
     // 申请集合通信操作的 Device 内存
     ACLCHECK(aclrtMalloc(&sendBuf, mallocSize, ACL_MEM_MALLOC_HUGE_ONLY));
@@ -62,7 +62,7 @@ int Sample(void *arg)
     // 申请 Host 内存用于存放输入数据，并将内容初始化为 DeviceId
     ACLCHECK(aclrtMallocHost(&hostBuf, mallocSize));
     float *tmpHostBuf = static_cast<float *>(hostBuf);
-    for (uint32_t i = 0; i < count; ++i) {
+    for (uint64_t i = 0; i < count; ++i) {
         tmpHostBuf[i] = static_cast<float>(device);
     }
     // 将 Host 侧输入数据拷贝到 Device 侧
@@ -94,7 +94,7 @@ int Sample(void *arg)
     ACLCHECK(aclrtMemcpy(resultHostBuf, mallocSize, recvBuf, mallocSize, ACL_MEMCPY_DEVICE_TO_HOST));
     float *tmpResultBuf = static_cast<float *>(resultHostBuf);
     std::cout << "rankId: " << device << ", output: [";
-    for (uint32_t i = 0; i < count; ++i) {
+    for (uint64_t i = 0; i < count; ++i) {
         std::cout << " " << tmpResultBuf[i];
     }
     std::cout << " ]" << std::endl;
@@ -119,7 +119,7 @@ int main()
     ACLCHECK(aclrtGetDeviceCount(&devCount));
     std::cout << "Found " << devCount << " NPU device(s) available" << std::endl;
 
-    int rootRank = 0;
+    int32_t rootRank = 0;
     ACLCHECK(aclrtSetDevice(rootRank));
     // 生成 Root 节点信息，各线程使用同一份 RootInfo
     void *rootInfoBuf = nullptr;
