+++ examples/02_collectives/03_allgather/main.cc
@@ -38,7 +38,7 @@
 
 struct ThreadContext {
     HcclRootInfo *rootInfo;
-    int32_t device;
+    uint32_t device;
     uint32_t devCount;
 };
 
@@ -47,14 +47,14 @@ int Sample(void *arg)
     ThreadContext *ctx = (ThreadContext *)arg;
     void *sendBuf = nullptr;
     void *recvBuf = nullptr;
-    int32_t device = ctx->device;
+    uint32_t device = ctx->device;
     uint64_t sendCount = 1U;
     uint64_t recvCount = ctx->devCount;
     size_t sendSize = sendCount * sizeof(float);
     size_t recvSize = recvCount * sizeof(float);
 
     // 设置当前线程操作的设备
-    ACLCHECK(aclrtSetDevice(device));
+    ACLCHECK(aclrtSetDevice(static_cast<int32_t>(device)));
 
     // 申请集合通信操作的 Device 内存
     ACLCHECK(aclrtMalloc(&sendBuf, sendSize, ACL_MEM_MALLOC_HUGE_ONLY));
@@ -64,7 +64,7 @@ int Sample(void *arg)
     void *hostBuf = nullptr;
     ACLCHECK(aclrtMallocHost(&hostBuf, sendSize));
     float *tmpHostBuff = static_cast<float *>(hostBuf);
-    for (uint32_t i = 0; i < sendCount; ++i) {
+    for (uint64_t i = 0; i < sendCount; ++i) {
         tmpHostBuff[i] = static_cast<float>(device);
     }
     // 将 Host 侧输入数据拷贝到 Device 侧
@@ -115,7 +115,7 @@ int main()
     ACLCHECK(aclrtGetDeviceCount(&devCount));
     std::cout << "Found " << devCount << " NPU device(s) available" << std::endl;
 
-    int rootRank = 0;
+    int32_t rootRank = 0;
     ACLCHECK(aclrtSetDevice(rootRank));
     // 生成 Root 节点信息，各线程使用同一份 RootInfo
     void *rootInfoBuf = nullptr;
